solved some compiler warnings that are now visible in arduino 1.0. Found a couple of unused variables, that I commented. Tried to solve the program memory warning message, and failed.

+++ Marlin/MarlinSerial.h
@@ -120,7 +120,7 @@ class MarlinSerial //: public Stream
 
     FORCE_INLINE void print(const String &s)
     {
-      for (int i = 0; i < s.length(); i++) {
+      for (int i = 0; i < (int)s.length(); i++) {
         write(s[i]);
       }
     }
